Moved custom modifier selectiveBorder from the theme package to the new modifier package, where all custom modifiers will now be placed. Moved animation of flipping icons of the MNXExpandableCard and the MNXDropDownMenu to the new modifier flipScale.
